[FIX] sale_project_forecast: fix the 'resource_id' change on update of task
Before this commit, When selecting a project and task within that project,
the assignees of that task did not automatically assign their resource_id
if SOL is set on the slot.

After this commit, now when a user selects a project and a task within that
project, the assigness of that task is now automatically assigned to the
resource_id even though SOL is set on a slot.

task-2857220

X-original-commit: 2bc45c1d4fb8dbd39d55e93fc0415257517cbb1d
Part-of: odoo/enterprise#28403