[FIX] planning_holiday: fix domain processed to only fetch shifts impacted by leaves
Before this commit, when a leave is validated, the allocated hours and
the working days fields for the shifts in which the planned dates have
the leave in the interval have to be recomputed. However, the domain
processed includes too much shifts and so the recompute will take more
time than expected.

This commit fixes the issue by ensuring the domain processed includes
only the shifts impacting by the leaves added or updated. This commit
also improves the performances to slow the less as possible the
validation of leaves in Time Off app.

Performance improvements:
========================
The time measured, to show if the fix improves the performance, is made
when the user valids a leave in the Time Off app.
Also, this measure is done in a db with just the demo data installed.

In a db with all apps:
- without the fix: it took 1440 ms
- with the fix: it takes 476 ms

In a db with just `planning_holidays` module installed:
- without the fix: it took 310 ms
- with the fix: it takes 80 ms

Related PR: #24619

closes odoo/enterprise#27473

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Co-authored-by: Kevin Baptiste <kba@odoo.com>