[FIX] documents_spreadsheet: Format LIST values that correspond to dates
.. and datetime.

Currently, formulas `LIST` yield the server value of fields date and
datetime, which are strings. As such, the o_spreadsheet lib will tag the
cell with such content as a string; which prevents  a proper formatting.

To solve this, we now convert the server values to a number for values
realted to date and datetime fields.

Task 2857297

closes odoo/enterprise#27489

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>