[FIX] sale_subscription: portal view was wrongly marking subscription as purchase orders
Before this commit, portal view of subscription was displaying a wrong breadcrumb:
"Subscription > Purchase Orders > S0123"

The issue was that the sale.order record was referenced by the 'order' key in
the qweb rendering options.

The 'order' name was already used by purchase order override of the portal template.

Taskid: 2855699
X-original-commit: ceaeb5e6c33a2d40f9037ae0b64737fb57e4f203
Part-of: odoo/enterprise#28509