[IMP] quality_control: add operation quality check type
The previous naming of the quality check type was quite confusing. The
'operation' type actually created a check by product and not a single
one by operation. The 'product' type created a check by Lot/SN. Those
were renamed to be more consistent. The old 'operation' type is now
'product' and the old 'product' type is now 'Quantity' (move_line).
On top of that, a new 'operation' type was added. This applies on the
operation, with one single quality check by picking.

Task-id: 2614033
Part-of: odoo/enterprise#27602