[FIX] l10n_pe_edi: handle errors reported in CDR
SUNAT has two ways of reporting an error when posting or cancelling
an invoice:
- either with a SOAP fault, when the document could not be processed
- or through the <ResponseCode> XML node in the CDR.

see section 4.1 of the SUNAT programmer’s manual

At the moment, we only handle the errors raised via SOAP faults.
As a result, when SUNAT reports an error via the ResponseCode node of
the CDR, we ignore the error.

With this fix, we no longer ignore those errors.

opw-2725105

closes odoo/enterprise#29481

X-original-commit: b398846a4f3aa602165f706e7c1d692d552a5b4e
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>