[FIX] industry_fsm_stock: fix the access error on task assignment
Before this commit, When the user has Project - user access and
Sales- Own documents only right then trying to assign him a task
created from SO gives access error due to insufficient right to
read the update the SO based on the task assignee.

So in this commit, Check the SO with a super user to assign the
SO to the task's assignee.

task-2858336

closes odoo/enterprise#31036

X-original-commit: f591c05821b928dcd76aaa879ef4e18acbb113d7
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>