[REM] l10n_de_pos_cert: remove incompatible library code
The `letterRendering` option was removed from the `html2canvas` library
in recent versions. Since the library is getting updated, we can get rid
of code that was only there to check whether that option should be
active or not (which was in order to cover an issue with the old
workings of the library).

closes odoo/enterprise#27527

Related: odoo/odoo#91773
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>