[FIX] account_winbooks_import: parent tax imported after child
The name of taxes is computed by merging the name of all
parent taxes together, based on the code:
```py
tax_name = " ".join([treelib[x] for x in [rec.get('TREELEVEL')[:i] for i in range(2, len(rec.get('TREELEVEL')) + 1, 2)]])
```

This leads to an exception when the parent tax is after its childs
within the files holding the taxes.

```
  File "/home/odoo/src/enterprise/saas-15.1/account_winbooks_import/wizard/import_wizard.py", line 700, in import_winbooks_file
    vatcode_data = self.import_vat(file_dir, vatfile, account_central)
  File "/home/odoo/src/enterprise/saas-15.1/account_winbooks_import/wizard/import_wizard.py", line 599, in import_vat
    tax_name = " ".join([treelib[x] for x in [rec.get('TREELEVEL')[:i] for i in range(2, len(rec.get('TREELEVEL')) + 1, 2)]])
  File "/home/odoo/src/enterprise/saas-15.1/account_winbooks_import/wizard/import_wizard.py", line 599, in <listcomp>
    tax_name = " ".join([treelib[x] for x in [rec.get('TREELEVEL')[:i] for i in range(2, len(rec.get('TREELEVEL')) + 1, 2)]])
Exception

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "/home/odoo/src/odoo/saas-15.1/odoo/http.py", line 644, in _handle_exception
    return super(JsonRequest, self)._handle_exception(exception)
  File "/home/odoo/src/odoo/saas-15.1/odoo/http.py", line 302, in _handle_exception
    raise exception.with_traceback(None) from new_cause
KeyError: '23040203'
```

In the winbooks import files of the customer, in the support ticket,
the parent `23040203` is clearly after its childs
`2304020301`, `2304020302`, `2304020303`:
```py
In [7]: from dbfread import DBF; [r.get('TREELEVEL') for r in DBF('PROFILE_codevat.dbf', encoding='latin').records if '23040203' in r.get('TREELEVEL')]
Out[7]: ['2304020301', '2304020302', '2304020303', '23040203']
```

The goal of this revision is to sort the taxes record on their
`treelevel` length, so parents are always before their childs.

opw-2716479

closes odoo/enterprise#27642

X-original-commit: 3327fa5a84ec138f92effebc582d7b36402e409b
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>