[FIX] account_followup: fix followup mix multi company
When executing cron followup task, in a multi company environment, if one same partner is set on multiple invoices (requiring a followup) from different companies the followup line is computed only once for the first company.
Followup reports for the invoices of the other companies are the generated from the followup report line of the first company)

To reproduce :
- Create 2 companies
- Create one autoexecute followup level sending an email for each company with a different message
- Create one invoice triggering the followup report on each company with the same customers
- Set both companies on allowed companies for the "System" user (id 1)
- Run "Account Report Followup; Execute followup" manually
Result : both emails are created using the template of one company
Attended result: each email should be created using the template of the company it is originating from

This commit adds a dependency on the compute method for followup_level to force the update when the company is changed in the context.

closes odoo/enterprise#27640

X-original-commit: 9ba90678150e58b6b57d32b3b6711aa3ebcadcf8
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>