[FIX] l10n_uk: use correct tax data for HMRC
When using a fiscal position along with the l10n module for the same
country, the tax report of that foreign country is also available
in the tax report menu.
However, they are ordered by country_id and if the other tax report
is before the UK tax report, _get_options() will say that the tax
lines of the foreign country are the ones to send to the HMRC.
This fix forces the use of the UK tax report as a reference for
which lines to use.

opw-2845739

closes odoo/enterprise#27634

X-original-commit: 1fa70b017e35c12c8e293438bfc9eaa3993017ce
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>