[FIX] web_dashboard: dashboard determines sub views sample mode
Before this commit, after the dashboard view would leave the
sample mode, the sub views would still be in sample mode. This was due
to the presense of a key "useSampleModel" in the globalState of the new
sub views to be instantiated after the dashboard update.
This commit remove that key at an appropriate time.

Forward-port-of: 9154196
Part-of: odoo/enterprise#27589