[FIX] web_dashboard: sub views do not use dashboard search defaults
In legacy views, the search default keys are not removed from the
action context after they are used. After a switch view from a legacy
view with search defaults to a dashboard view, those search defaults
can still be found in the dashboard view contexts because the dashboard
view reuse the search state of the legacy view. This lead the dashboard
sub views to process themselves the search defaults, so that their own
domains can become different from the dashboard domain.
Here we remove the search default keys from the contexts passed to sub
views. In that way, the dashboard view has full control on the sub view
domains as it was intended.

Forward-port-of: bd1600b
Part-of: odoo/enterprise#27589