[FIX] web_dashboard: keep states of sub views
The dashboard view do not reuse its sub views when it rerenders. New sub
views are created each time. Unfortunately, a new sub view did not reuse
the state of the previous corresponding view when the dashboard would
have its props updated, so that the user actions on that sub view would
be forgotten. For example, it the user would change the mode of a graph
sub view  from "bar" to "line" and then would change the dashboard view
domain via the control panel, the new graph sub view would be in "bar"
mode instead of "line" mode.

Here we make sure that a sub view transmits its state to the next
corresponding sub view.

closes odoo/enterprise#27589

Forward-port-of: da099be
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine <dam@odoo.com>