[FIX] helpdesk: align the kanban state to the right
Currently, when no `user_id` is set on the `helpdesk.ticket`, the kanban state
takes up the current space and as a result there is some blank space on the
right side of the kanban state.

So this commit applies the condition on `user_id` field so that
the user avatar will not take up space on the kanban card if the
user avatar is not available.

task-2774812

closes odoo/enterprise#27592

X-original-commit: 52a720d
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>