[FIX] hr_payroll : no decimal in wage for payslip workdays lines
Step to reproduce:
	-install payslip
	-set on an employee contract a wage with decimal value
	-set attendance for this employee (if he did not already
	have some)
	-create a payslip for this employee with the modified contract

Current behavior:
The wage in the "worked days & inputs" and "salary computation" does not
have the decimal values of the contract wage (00 instead)

Expected behavior:
The wage in the "worked days & inputs" and "salary computation" have the
decimal values of the contract wage

Explanation
The worked_days had their amount computed based on the deprecated
normal_wage value of the payslip that is an integer and does not
take into account the decimals of the contract_wage. To solve the
issue we replace the normal_wage with the contract_wage. We also overwrite
the _compute_amount function of hr_contract_salary_payroll to make
it depens on the contract wage_on_signature.

opw-2845145

closes odoo/enterprise#27756

X-original-commit: 9274e025a1ad35f3b7422f580feff68b0579a2b6
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>