[FIX] appointment: fix disappearing message intro
Before, each time we wrote on an appointment type we set to False
the message_intro if we didn't update it.
We now remove this part as it leads to side effect on performance.

The additional queries are linked to the "res.lang" model.
Indeed with the previous behaviour, the field was automatically set
to False. This led to ignore translation feature for the field each
time we wrote on records.

X-original-commit 6704fc6b537a3f5e90af1d4e60793774c870d401

task-2858271

closes odoo/enterprise#28688

X-original-commit: 09ce989b73ea224ddd796ed0cf41d5444d0a7007
Signed-off-by: Hennecart Jérémy (jeh) <jeh@odoo.com>
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>