[IMP] documents: prepare activity menu tests to wowl env
This PR prepares the ground for the one introducing the new environment in the discuss app.
Indeed, the start helper won't be able to handle widget anymore. To solve this issues,
we're now adding the activityMenu widget to the systrayMenu.Items array, this will be
mapped by the createWebClient helper to component and added to the new systray item
registry.

task-2582313

closes odoo/enterprise#27614

Related: odoo/odoo#91953
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>