[IMP] stock_barcode: picking type scan options
This commit adds new fields for the `stock.picking.type` to set some
scan as mandatory. The purpose is to force picker to scan some type of
barcode at a given step in the operation to reduce the handling errors.

The operation will always ask to scan barcodes in this order:
  - Scan the source location;
  - Scan the product;
  - Scan the lot/serial number;
  - Scan the package;
  - Scan the destination location.

Only if the setting for the corresponding field is set as `mandatory`
and if it's relevant (it won't ask to scan the source location for a
receipt or if multilocation is disabled for example).

If the product's scan is mandatory:
  - A line can't be selected if its product wasn't scanned yet (except
    if the product has no barcode);
  - The edit/+1 buttons can't be used if the product wasn't scanned yet;
  - A tracking number can't be scanned until a product was.

If the tracking number is mandatory:
  - A line for a tracked product can't be edited while its serial/lot
    number wasn't scanned.

If the source location is mandatory:
  - The user must scan the location's barcode before to be able to scan
    anything else;
  - The user can't change the location with the locations' list.

If the source location is mandatory:
  - Once the user has started to scan a product, they must scan the
    destination location before to be able to scan another product;
  - The user must scan the location's barcode before to be able to
    change the page or validate the transfer;
  - The user can't change the location with the locations' list.

If the package is mandatory:
  - Once the user has started to scan a product, they must scan a
    package or do a put in pack before to be able to scan anoter product
    or the location.

The barcode's instruction is reworked to be more specific depending of
the configuration and of what was already scanned or not.

Also, if the user scans something else than what is expected while it's
mandatory, a prompt will be displayed to ask to scan the right barcode.

task-2666478

Part-of: odoo/enterprise#27547