[FIX] sale_subscription: update next_invoice_date when needed.
Before this commit, the next_invoice_date was always incremented when the
subscription invoice was created.

It caused issues when the payment_mode  is 'success_payment'
and the payment fails because the token is empty or expired.

Step to reproduce:
* Create a subscription, assign a token, invoice the first period.
* Delete the token before the next invoice and run the invoicing cron
* The new invoice is created and the next_invoice_date is incremented.
* The automatic payment fails but the next_invoice_date is not set back to
  the previous value.
* Manual intervention is needed to allow to recreate the invoice when the new token
 is saved.

Now, the next_invoice_date is not automatically incremented for success_payment
payment_mode. It is done in 'reconcile_pending_transaction' as it was done before.

task-2824236

closes odoo/enterprise#27750

X-original-commit: 1cf72b6ba532dcb9027b6c0ddb34e53a384224cf
Signed-off-by: Arnaud Joset <arj@odoo.com>