[IMP] *.*: update the tooltips in various modules
The goal of this task is to review the copywriting of the tooltips
because some of them are not correct in English and others are not
valid anymore. This is also a good opportunity to make an inventory
of the tooltips we have and to add some that could be missing.

task-2860991

X-original-commit: 0022d758158860fdd0f2a1eb044a9f887db68fb1
Part-of: odoo/enterprise#32298
Co-authored-by: Mahendra Barad <mba@odoo.com>
Co-authored-by: Prakash Prajapati <ppr@odoo.com>