[IMP] documents_spreadsheet: remove quotes from IDs on insertion
Task Description

When insterting a pivot/list in a spreadsheed, we got formula like
'=PIVOT("1","foo","true","record_id","1") and "=LIST("1","2","foo"),
where the IDs are encapsulated between quotes.

The aim of this task is to remove these quotes in the formulas when the
values is a number (pivot ID, list ID, record ID, ...) as they are not
necessary anymore and can make the formula less-readable.

To be consistent, the quote has also been removed from the autofill of
pivots/lists, as well as the pivot template.

Related Task

-task-2832354

closes odoo/enterprise#27538

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>