[FIX] account_online_synchronization : prevent multiple journals on bank sync
To reproduce
============

- on a online sync bank account
- create a new journal (from type to create)

when saving, a Validation Error is raised "You cannot have two journals associated with the same Online Account.".
But when reloading the page (F5), we found out that the journal was added to the record.

this bug can be reproduced an other way, by duplicating a journal that is linked to an online account, when checking
this online account we find that it has two journals (the original and the copy).

Purpose
=======

For the first case, the issue is caused by not setting a constraint on `account_online_account_id` in `account_journal`
to check if the online account already has a journal or not.

For the second case, an online account should have only one journal, so when duplicating a journal we must not copy the `account_online_account_id`.

Specification
=============

to solve the issue we add constraint on `account_online_account_id` and give it the attribute `copy=False`

opw-2851216

closes odoo/enterprise#27705

X-original-commit: bdcc9975de7e5908614e0f74713dccaa7f83c049
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: abla001 <abla@odoo.com>