[FIX] documents_spreadsheet: Import Odoo link cells from data
It was not possible to load a cell containing an Odoo link from the data
because its constructor was using getters declared in a UI plugin. When
importing the data, such plugins are not yet loaded and should not be
used to create core data anyway.

In order to import such cells, there are to requirements:
- the cells constructors need to be loaded before we start the import
process
- the constructors need to access the env services (specifically `menu`)

To solve this issue, we introduce a new service `spreadsheetLinkMenuCell`
which sole purpose is to load the different odoo link constructors to
the cell factory, before we create an instance of `Model`, hence before
the import process.

Task 2860257

closes odoo/enterprise#27914

X-original-commit: aed0dd737b00305ece4aab2d14bec7a44dfa82f9
Related: odoo/odoo#92577
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>