[IMP] voip: adapt dialig panel tests for wowl env
This PR prepares the ground for the one introducing the wowlEnv in the discuss app.
Indeed, the start helper won't return widgets anymore. This means we need to adapt those
tests to stop calling widget methods and rely on a more realistic behavior. The voip dialing
panel tests have been adapted to ease the wowlEnv integration.

task-2582313

closes odoo/enterprise#27720

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>