[REF] *: remove kanban .includes
The goal of this commit is to remove various .includes done in the js
codebase. This makes the code more robust, and will help us soon when we
will replace the kanban views with the new implementation.

Many kanbanrecord .includes can be removed by using the new
'action'/'type' attributes on kanban views. Some other had to be
reimplemented using the standard sub view approach (with a js_class
attribute).

closes odoo/enterprise#27747

Related: odoo/odoo#92071
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>