[FIX] partner_commission: set referrer_id on the invoice created from SO
The Issue:
After creating a sale order, then create a new invoice with a down payment (percentage option), the invoice doesn't have the referrer_id even though that it was set on the sale order.

The fix
Adding the referrer_id to the invoice created from a sale order.

opw-2820420

closes odoo/enterprise#27847

X-original-commit: 00b24f13af6864a356404868c691bbebf49f1e58
Signed-off-by: Arnaud Joset <arj@odoo.com>