[FIX] hr_contract_salary: Fix email in simulation link.
Change the default recipient of the simulation link to be the employee's
work email address, instead of the private email address.

Currently the private email of employees is used for this but this doesn't
work because mail.compose.message does not allow allow private partners
to be configured as recipients. Secondly, it also just makes sense to send
work-related email to the work email address.

task-2790044

closes odoo/enterprise#27808

Related: odoo/odoo#95655
Signed-off-by: Kevin Baptiste <kba@odoo.com>