[FIX] quality_control: delete checks on SM cancellation
When cancelling a SM, it also cancels all quality checks of the
associated picking. This is not accurate enough.

To reproduce the issue:
1. In Settings, enable "Multi-Step Routes"
2. Edit the warehouse:
    - Incoming Shipments: 2-steps
3. Create two storable products P1 and P2
4. For each product, add a quality control point:
    - Operation: Internal Transfer
5. Process a receipt (to the input location) with:
    - 1 x P1
    - 1 x P2
6. Process a second receipt (still to the input location) with:
    - 1 x P1
7. Open the internal transfer for the received produits
8. Open the quality checks

Error: There is only one quality check (for P1), the QC of P2 is missing

When processing the second receipt, the module creates an internal
transfer to move the received P1 from input to stock location. However,
because there is already a similar SM (from the first receipt), both SM
are merged and the second one is cancelled:
https://github.com/odoo/odoo/blob/c08ebc3a24f2326bfdf5a98b5533a501ade57c04/addons/stock/models/stock_move.py#L768-L772
As a result, in `quality_control`, we delete all quality checks (then a
new one will be created for the merged SM)

OPW-2837078

closes odoo/enterprise#28462

X-original-commit: 29cbc2917d293d8befebde96e541fb0ceae048d5
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>