[FIX] sale_subscription: let sale compute the payment.transaction reference
Before this commit, the payment.transaction reference was manually calculated
but the sale module has everything to let the reference be computed with the
order name automatically. When sale.subscription records existed, the record
name was computed and could not be used as a reference. The 'code' SUB00X was
what we use now as the name S00XX.
This commit will prevent weird behavior where subscription transaction created
by token payment would have different reference than the rest of the other SO's.

Taskid: 2855699
X-original-commit: f7583b44dce739783f009c922e7ab881983800ca
Part-of: odoo/enterprise#28509