[FIX] helpdesk: fix the traceback on coupon wizard
Currently, On creating of a coupon from help-desk,
and after genrating a coupon, on clicking cancel >>
generate a traceback due to invalid action on button.

so in this commit, fixes the issue by replacing the cancel
with type special and also make the expire date as readonly
as it's just wizard to send the coupon.

task-2857229

closes odoo/enterprise#28708

X-original-commit: 4daf5a23e8208ede7e1e272586c3ecd664a62cba
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>