[FIX] account_reports: correct date selection in reports
Steps to reproduce:
(on a new database without demp)
- create an invoice of 1000$ for every month from November 2021 to April 2022
- In reporting > Balance Sheet
- Select "End Of Last Month"

Issue:
The "Previous Year Unallocated Earnings" will be incorrect != 2000 (1000 December + 1000 November)

Cause:
The issue is not from "Previous Year Unallocated Earnings" per se but from from one of its subcomponent: the "Current Year Earnings"
PREV_YEAR_EARNINGS =
    -sum                         : from_beginning
    + ALLOCATED_EARNINGS        : from_beginning
    - CURR_YEAR_EARNINGS_ALLOC    : from_fiscalyear
    - CURR_YEAR_EARNINGS_PNL    : normal            -> Problematic one
        [OPINC + OIN - COS - EXP - DEP]

The CURR_YEAR_EARNINGS_PNL or NEP (Net Profit) is computed for the month of April 2022 (04/01/22 - 04/30/222) instead of from the beginning of the fiscal year (01/01/22-30/04/22).
This happens only when you select a "Last of [month quarter]" option (quarter if in the third or fourth one). If you select "Today" or a "Custom End Date", it won't cause the issue.

Solution:
Force a "custom" option value to compute it correctly

opw-2821529

closes odoo/enterprise#28046

X-original-commit: 062639067160235e26077c5c66767662003be8e3
Signed-off-by: yosa-odoo <yosa@odoo.com>
Signed-off-by: Olivier Colson <oco@odoo.com>