[IMP] sale_stock_renting: speed-up _compute_is_available
Replace for loop with an if condition by a call to filtered
followed by a for loop. This sets the _prefetch_ids of
rental.order_line_id to only the sale.rental.schedule
records that pass the condition. This avoids
useless computations when getting
the sale.order.line.virtual_available_at_date computed
field.

closes odoo/enterprise#36506

X-original-commit: 1cd2502b90763079b28bf8137beafc8364fc33cf
Signed-off-by: Van Delft Aurélien (avd) <avd@odoo.com>