[IMP] sale_stock_renting: speed-up _compute_is_available
Replace for loop with an if condition by a call to filtered
followed by a for loop. This sets the _prefetch_ids of
rental.order_line_id to only the sale.rental.schedule
records that pass the condition. This avoids
useless computations when getting
the sale.order.line.virtual_available_at_date computed
field.

closes odoo/enterprise#33552

X-original-commit: 832a6709889af2ba7a6a66115d7bcfd4a98e2b4e
Signed-off-by: Van Delft Aurélien (avd) <avd@odoo.com>