[IMP] *: make orders give them recurrence to the tasks they create
* =	industry_fsm_sale_subscription,
	project_enterprise,
	project_sale_subscription,
	sale_subscription

On SO confirmation:
- If all of the following conditions are met,
set the end_date of the Subscription
as the repeat_until date of the Task:

The Product is recurrent and creates a Task,
The Task Template is recurrent,
The end_date field is set on the Subscription.

- If all of the following conditions are met,
apply the Subscription's recurrence to the Task's:

The Product is recurrent and creates a Task,
There is no Task Template on the Product.

- If the Subscription is closed, set the recurring_task field
of the related Task to false.

related PR: https://github.com/odoo/odoo/pull/92696
task-2868306

closes odoo/enterprise#27953

Related: odoo/odoo#92696
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>