[IMP] account_bank_statement_import_camt: clearer error message
When importing a CAMT file, we are filtering statements using the
journal currency. However, if a wrong currency is set on the journal, no
statements are imported and an unclear error message is raised in
_check_parsed_data. We should adapt so the account number is displayed
correctly in that generic error message instead of False, and we should
also add a new specific message so the user knows he has to check the
journal currency in order to possibly fix his issue.

closes odoo/enterprise#28953

X-original-commit: c3aa598a6a4dcef79dba9d441ed1d7c3f6a97349
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>