[FIX] website_calendar: exclude appointment info from sitemap
Before this commit if `/info` is added to the url of an online
appointment (e.g. `/calendar/appointment_type/info`), without
specifying the arguments, it throws a server error. This is
particularly problematic as in v14 the pattern is added to the
site map. This causes problems with search engines and crawlers
and prevents successful crawling of the website.

We can prevent unsuccessful crawls by excluding this URLs
from the site map. (See discussion in:
https://github.com/odoo/enterprise/pull/27894)

opw-2856276

closes odoo/enterprise#28647

X-original-commit: f54e0e7ccac3fc3744f02075984936c3a98ba99a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>