[IMP] account_reports_cash_basis: remove trailing comma
A template inherit for cash basis is adding a comma after the
'included posted entries' option, even when the additional part after
it isn't added.
This will to remove that trailing comma in case cash basis isn't
enabled on the report.

This also move this part to the button itself, instead of the option
in the list as it doesn't make sense.

Task id ##2838182

closes odoo/enterprise#27840

Related: odoo/upgrade#3560
Related: odoo/odoo#92438
Signed-off-by: Quentin De Paoli <qdp@odoo.com>