[IMP] social: improve calendar view
PURPOSE

This commit allows one to drag and drop the draft post
on the calendar view. It also color the post depending
on the utm campaign instead of its state. The draft state
are hatched to easily distinguish them from others. One
can no longer create or move a draft or scheduled post
in the past.

LINKS

Task-2860008

closes odoo/enterprise#27931

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>