[FIX] l10n_de_reports: Adapt length of accounts for datev Export
Before this commit, we always padded to 8 digits the `konto` and 9 digits the `gegentokonto`.

Now, we pad the `gegentokonto` to the maximum of the length of both system parameters (default to 9 if a parameter is missing).
`konto` is padded to the same length minus 1.

We also take care of the fact that the length shouldn't be outside 5 and 9.

task-2869757

closes odoo/enterprise#28795

X-original-commit: 1f9b22c7ba2bae97c42ef8f9e45f29a9ad44ba89
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>