[FIX] l10n_de_pos_cert: use product name as line text
Before this commit, template `dsfinvk_cash_point_closing_template` used
`full_product_name` for the `text` field for order lines. When a sale
order is imported in the POS session, this field includes the product
description as the full product name by default. This description
might include line breaks (i.e. `\n`) which causes an error in
`ast.literal_eval` when processing the POS session. It might also be
too long which doesn't comply with Fiskaly API (which specifies it to
be 255 characters or less)

To fix these problems, we can use product name as the text field for
order lines in this report (the same way it was done in v13)

opw-2817707

closes odoo/enterprise#27987

X-original-commit: a84e7548676d33ae64f59f661cd633515e047b17
Signed-off-by: Trinh Jacky (trj) <trj@odoo.com>