[FIX] sign: no translations in frontend
In commit 00e6b87c9d16af1bfb3e63b9b857fbaad90e620a the deletion of
some code created a bug in the sign app where the language is set
in the html_data and the frontend translations were hence broken
(these translations were introduced by 7d59eb832d727522fc3b4639634c377e66ce3c59).
In the stable versions this commit was revert but on master we
find a solution in the sign module.By creating the
get_frontend_session_info_sign we add the lang to the user_context.

We also remove the deprecated xml line of the first fix commit
7d59eb832d727522fc3b4639634c377e66ce3c59 that are now irelevant
as we don't use the html_data anymore.

opw-2777044

closes odoo/enterprise#28037

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>