[FIX] documents_spreadsheet: Correctly load data in PivotDialog
In certain conditions, the pivot data were not correctly fetched from the
`pivotModel`when loading the `PivotDialog`to re-insert a cell.

Specifically, the issue arises when  at least one groupBy is a
relational field which is not set for some records. (e.g. user_id not
set on res.partner records).

The problem was that when parsing the groupBy value in `parsePivotFormulaFieldValue`,
the code only worked if the input groupValue was always a string.

This commit ensures that `parsePivotFormulaFieldValue` properly handles
all types of inputs, therefore fixing the aforementionned symptom.

Task 2870790

closes odoo/enterprise#28285

X-original-commit: 9559e53e01b0b99c3ee1a6f0cc1b53ed1c5e7385
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>