[IMP] {website_}appointment: add an operator selection view and flow
RATIONALE

In order to allow the customers to see appropriate availabilities, a
new view is created as a default appointment_type view: an operator
selection view, with cards of the available users, that one can pick
to get to the previous 'calendar time selection' view. This view is
only shown if appropriate. (cfr below)

OPERATOR SELECTION (only when WEBSITE is installed)

A new template appointment_select_operator introduced to support this new view.
The default route appointment_type_page will automatically renders it, instead
of the default appointment_type_info, when these conditions are met:
- Avatars must be activated on the appointment_type.
- There is more than one suggested user (after filtering).
- No staff_user_id is set in the route parameters. This is only considered
if that id is among the suggested users. It will be used as selected user.
- The assign method of the appointment type is 'chosen'.
The view is not associated to a new route in order to avoid redirections when
the condition is met, as well as duplicated computations and code.

The view consists of selectable cards of all the possible staff users, with
their picture, their linked partner's function and their website_description
if any. Function and description are editable in front-end.
-> Therefore, website_partner module is now part of the depencencies of the
module website_appointment. However, it is not necessary for the auto-install.
It will be installed if other dependencies are, at the auto-install.

To implement this default behaviour, appointment_type_page route is
restructured using methods that can be overriden in website_appointment or
other modules, notably _get_appointment_type_page_view that is used to pick
the template rendered depending on the condition of the operator view.

The new structure allowing easy overrides goes as follow:
- A method will prepare page values (useful to choose the template or to render
them) used further.
- A method selects which page to render and returns that rendering.
(In the appointment info view, which is the default view, the slots and other
values are computed and the template is rendered.)

staff_user_id in url is used to transmit the selected user to the appointment
type page route. This parameter is propagated through the different views to
avoid a redundant selection in case of an error (slot already taken / user not
available) -> The user will be redirected to the time selection view with the
correct user preselected. If the parameter is used for 'chosen' a_t not meeting
the requirements to use the selection, it will be pre-selected in the dropdown.

Task-2496629

Part-of: odoo/enterprise#24744