[IMP] {website_}appointment: redesign and move navbar to website_appointment
The styling is moved to website_appointment as we now remove chevrons from
the customer preview. They are completely redesigned as a part of the new
operator selection view: operator, time, details and booked chevrons.

The new template appointment_progress_bar is easy to call and work on the
following parameters: step, users_possible, appointement_type.
The integer step is used in order to add additional steps later easily.
These are the only parameters to set when calling this chevron template.
Time and Operator chevrons are made clickable and an hovering effect is added
for passed chevrons. Colors match the theme ones.

Use the same 4 chevrons on the following front end views : operator, time,
details. The operator chevron is only shown if that step is available.
The url parameters determine which template is used, so no staff_user_id
is kept when clicking the 'operator' chevron. The design of the navbar is
harmonized as well in terms of position and color. However, no chevrons are
displayed on the validation screen.

As we need to know how many available appointments there are in order to
display (or not) the 'all appointments' links in the views to go back to
the appointment selection, we add them in the page values.

Url parameter redundancy:

As the chevrons and errors / cancelations make the appointment process cyclic.
This makes it hard to deal with redundancy of parameters when selecting another
slot or user.
-> We could specify what parameters we want to keep but this makes the templates
longer and inheritance quite hard to read. -> Instead, we keep everything, when
coming back from error / cancelation , or going from  'Details' to 'Form'
using chevrons.
-> To prevent redundancy of date_time and duration, we delete them when selecting
a slot. This makes sure we only get the relevant one and 'cut' the loop that
keeps adding those parameters in the url. Unicity is assured.

Task-2496629

Part-of: odoo/enterprise#24744