[FIX] web_gantt: breaking test on Chrome 94+
Since Chrome 94, some element's sizing computation returns a
slightly different value (in the order of a fraction of a pixel). Sadly,
due to rounding, this difference has an impact on exact sizing
assertion.

This Chrome change impacted the "Connectors are displayed behind pills,
except on hover" test and more specifically the
`document.elementFromPoint" call (targetting the
`o_connector_stroke_hover_ease` path instead of the the actual
`o_connector_stroke`.

This commit fixes it in two ways:
- as the "hover" state is global to the whole connector, targetting the
  `o_connector` makes more sense than an inner element ; adapting
  the other tests for consistency.
- slightly adjust the position relative to other elements (the
  `o_gantt_pill` in this case)

closes odoo/enterprise#28215

X-original-commit: 41e65f25d1a07e2c9d6d3986dd29a03f1e38e6ea
Related: odoo/odoo#93235
Signed-off-by: Pierre Paridans (app) <app@odoo.com>
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>