[IMP] social*: add username in accounts and account owner in accounts/streams
This commit adds the username/handle (if available) in the social accounts (form
and list view), by introducing the `social_account_handle` field. Note that for
Twitter, this field will contain the same info as the `twitter_screen_name`
field, so this commit removes the `twitter_screen_name` field.

Apart from that, this commit also adds owner (create_uid) in the tree view of
social accounts and social streams to easily determine which account belongs to
whom.

taskID-2857845

closes odoo/enterprise#28067

Related: odoo/upgrade#3631
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>